printf_dev.h

+++ ft_printf_dev.h
@@ -6,7 +6,7 @@
 /*   By: alcierra <alcierra@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/31 16:54:08 by alcierra          #+#    #+#             */
-/*   Updated: 2021/12/08 15:35:40 by alcierra         ###   ########.fr       */
+/*   Updated: 2021/12/14 16:03:28 by alcierra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,23 +25,23 @@ int			ft_strlen(char *str);
 int			ft_putstr(char *str);
 
 // ft_putnbr.c
-int			ft_putnbr(long long num);
-int			ft_putnbr_neg_fd(long long num, int fd);
-int			ft_putnbr_fd(long long num, int fd);
-int			ft_putnbr_usign_fd(unsigned long long num, int fd);
+int			ft_putnbr(long num);
+int			ft_putnbr_neg_fd(long num, int fd);
+int			ft_putnbr_fd(long num, int fd);
+int			ft_putnbr_usign_fd(unsigned long num, int fd);
 
 // ft_putnbr_base.c
-int			ft_putnbr_base_neg_fd(long long num, char *base, int fd);
-int			ft_putnbr_base_fd(long long num, char *base, int fd);
-int			ft_putnbr_base_usign_fd(unsigned long long num, char *base, int fd);
+int			ft_putnbr_base_neg_fd(long  num, char *base, int fd);
+int			ft_putnbr_base_fd(long  num, char *base, int fd);
+int			ft_putnbr_base_usign_fd(unsigned long num, char *base, int fd);
 
 // ft_putpointer.c
-int			ft_putptr_upper(unsigned long long num, int fd);
-int			ft_putptr_lower(unsigned long long num, int fd);
+int			ft_putptr_upper(unsigned long num, int fd);
+int			ft_putptr_lower(unsigned long num, int fd);
 
 // ft_dec_to_another
-char		*ft_dec_to_another(unsigned long long num, char *base);
-int			ft_get_base_digits(unsigned long long num, int base_len);
+char		*ft_dec_to_another(unsigned long num, char *base);
+int			ft_get_base_digits(unsigned long num, int base_len);
 
 // ft_number_operations.c
 long long	ft_get_highest_digit(long long num);
